Fixes #2 #3 😄

+++ index.js
@@ -4,6 +4,7 @@ const path = require('path');
 const rimraf = require('rimraf');
 const imagemin = require('imagemin');
 const imageminZopfli = require('imagemin-zopfli');
+const args = process.argv;
 
 const icons = {
   60: 'apple-touch',
@@ -37,12 +38,11 @@ function resizeImages(imageBuffer, size, cb) {
   // https://github.com/aheckmann/gm
   gm(imageBuffer)
     .resize(size, size)
-    .compress('Lossless')
     .write(
       path.resolve(__dirname, `./.tmp/${icons[size]}-icon-${size}x${size}.png`),
       err => {
         if (err) {
-          console.log('shit', err);
+          console.log(':( An error occured.', err);
         } else {
           console.log(
             `./${icons[size]}-icon-${size}x${size}.png ===> Resize complete 🎉`
@@ -53,47 +53,57 @@ function resizeImages(imageBuffer, size, cb) {
     );
 }
 
-fs.readFile(path.resolve(__dirname, 'sample.png'), (err, imageBuffer) => {
-  if (err) console.log('shit', err);
-  if (!fs.existsSync('.tmp')) fs.mkdirSync('.tmp');
-  if (fs.existsSync('icons')) rimraf('icons', () => {});
+// args[0] == node bin path
+// args[1] == this index.js path
+// args[2] == user-provided image source path
+if (fs.existsSync(args[2])) {
+  // Read image source from filesystem into a buffer
+  fs.readFile(path.resolve(args[2]), (err, imageBuffer) => {
+    if (err) console.log(':( An error occured.', err);
+    if (!fs.existsSync('.tmp')) fs.mkdirSync('.tmp');
+    if (fs.existsSync('icons')) rimraf('icons', () => {});
 
-  let resizePass = Object.keys(icons).reduce((promiseChain, size) => {
-    return promiseChain.then(
-      () =>
-        new Promise(resolve => {
-          resizeImages(imageBuffer, size, resolve);
-        })
-    );
-  }, Promise.resolve());
-
-  console.log('⏱  Resizing...');
-  resizePass.then(() => {
-    console.log('✅  Success!');
-    console.log('');
-    console.log('');
-
-    let optimizationPass = Object.keys(icons).reduce((promiseChain, size) => {
+    let resizePass = Object.keys(icons).reduce((promiseChain, size) => {
       return promiseChain.then(
         () =>
           new Promise(resolve => {
-            optimizeImages(size, resolve);
+            resizeImages(imageBuffer, size, resolve);
           })
       );
     }, Promise.resolve());
 
-    console.log(
-      '⏱  Optimizing... larger icons will take longer... use the fifth P!'
-    );
-    optimizationPass.then(() => {
-      console.log('✅  Success! Thanks for being patient.');
+    console.log('⏱  Resizing...');
+    resizePass.then(() => {
+      console.log('✅  Success!');
       console.log('');
       console.log('');
 
-      console.log('⏱  Cleaning Up...');
-      rimraf('.tmp', () => {
-        console.log('✅  Success!');
+      let optimizationPass = Object.keys(icons).reduce((promiseChain, size) => {
+        return promiseChain.then(
+          () =>
+            new Promise(resolve => {
+              optimizeImages(size, resolve);
+            })
+        );
+      }, Promise.resolve());
+
+      console.log(
+        '⏱  Optimizing... larger icons will take longer... use the fifth P!'
+      );
+      optimizationPass.then(() => {
+        console.log('✅  Success! Thanks for being patient.');
+        console.log('');
+        console.log('');
+
+        console.log('⏱  Cleaning Up...');
+        rimraf('.tmp', () => {
+          console.log('✅  Success!');
+        });
       });
     });
   });
-});
+} else {
+  throw Error(
+    'You must provide a source image to the `npm start` command. See README.md for details.'
+  );
+}
